MANIT: Fix collapsed layout
